Use pandoc's YAML meta-data block syntax to give titles to docfiles
These are useful when the filename is encoded gibberish, which is
necessary in turn to avoid generating filenames that include bad
characters ( / is forbidden on Unix, for example, and others are
forbidden on Windows, and others still cause all sorts of pain for
most shells).

But the HTML <title> element doesn't need to be present in the file
otherwise and pandoc generates an <h1> for it, so the .css file causes
those h1 elements not to display (and is otherwise a copy of what
pandoc would generate).