Fix #(FILE=...) format to be much more liberal
Filenames can take weird and wacky forms and rather than have a
regular expression trying to capture all of them via listing
explicit allowed characters, it's easy to express the idea by
excluding the special cases (that will have to be escaped with
backslashes).  The special chars are the parentheses, the space
character, the double quote (") and the backslash.  Newline should
possible also be handled.

This approach allows for Unicode in paths also.