Rename theory while to while_
In the long term, we may want to drop the Theory suffix from theories, which
would cause issues with whileTheory, as while is a keyword in Standard ML.

Less importantly, the HOL + SML parser does not like the fact that there is a
keyword in the list of ancestors, since already with the Theory syntax the
Theory suffix is not mentioned anymore.

Appending a keyword with _ seems to be a common solution (see PEP 8). An
additional upside is that if we go with snake case as the naming convention for
theories (as is mostly the case in CakeML), while_ would be valid (in contrast
to While).
Also, functor was already renamed to functor_.