HolSmt: handle BBLAST_PROVE exception
BBLAST_PROVE by default raises a Match exception when a counterexample is found.

However, the counterexample may be spurious due to BBLAST_PROVE assigning some
value to a Z3 variable, even though the Z3 variable may have a different (and
yet-to-be-known) value.

To fix this issue, we temporarily configure BBLAST_PROVE to not try generating
counterexamples (there's no point in doing so when replaying a Z3 proof) and
then catch the HOL_ERR exception as usual so we can continue trying other
proving strategies.