Prove alternative set_byte and word_of_bytes thms
If we know the byte being set is zero, then set_byte is equivalent to just a
bitwise-or of the byte (cast and shifted into place) and the original contents.
In particular, word_of_bytes can be shown to be just a bitwise-or of all the
bytes being inserted, assuming we're not in a funny overflow case.

This makes at least some downstream proofs about word_of_bytes much easier.