More work with BB convs
Main issue at the moment appears to be that traversal loops with the
MODEQ weakened relation. In particular, 7 rewrites to 7 MOD 4, say,
but then the calculation is not done, and instead the rule that says
MODEQ n (x MOD n) x fires and this takes us back where we started.