Replace many (Q.)prove instances with Theorem ..[local] in src (#1692)
* Replace many Q.prove instances with Theorem ..[local] in src

* Replace many prove instances with Theorem ..[local] in src

* Replace many prove + Term instances with Theorem ..[local] in src

- Manually factored out "option_not_empty"
- Added space after some : in Theorem, as the HOL + SML parser is unhappy
  without them

* "Fix" store_thm_at override in pred_set

Some proof was failing in listScript.sml. With the power of binary search and
magit, I was able to reduce the issue to IN_IMAGE'. Now this made no sense,
because IN_IMAGE' did not appear to be exported. But in fact, courtesy to the
store_thm_at override, which appears to have ignored the use of "local",
it *did* export a rewrite. Amazing!

I did consider renaming IN_IMAGE' to IN_IMAGE_OTHER, because in the latter case
it does not get exported. However, I think checking whether something is a local
theorem is cleaner.

* Remove some empty lines