Add new syntax for EL that makes it look like array-subscripting
The term

   EL n l

now prints as, and can be entered as

   l❲n❳

where this is a tight suffix so that something like

   f l❲n❳

is actually

   f (EL n l)

and to get EL n (f l) one would have to write/read:

   (f l)❲n❳

The brackets are U+2772 and U+2773 respectively.

The new syntax (and how to turn it off) is documented in the
DESCRIPTION manual.